build: enable -Wall for main component and fix all surfaced defects
The root CMakeLists passed -w project-wide, silencing every warning in
every component including our own main/. Enable warnings for main/ only
and fix every defect that surfaced.

Warning policy (CMakeLists.txt, main/CMakeLists.txt):
- Drop the blanket project-wide -w; compile the main component with
  -Wall -Wextra.
- Re-apply -w to all non-main components (ESP-IDF framework, managed
  LVGL/esp_lvgl_port, vendored components/*) via a post-project() loop.
  A blanket -w cannot be re-enabled per component because -w overrides a
  later -Wall regardless of order, so it must simply be absent from main.

Defects fixed in main/ (all -Wall/-Wextra clean, ESP-IDF promotes several
of these to errors via -Werror=format/-Werror=pointer-sign/-Werror=return-type):
- return-type: indicator_model.c, display_model.c and btn.c (two weak
  stubs) were int-returning functions that fell off the end; now return 0.
- weak/strong signature mismatch: _sensor_data_parse_handle was declared
  size_t in rp2040.c but ssize_t in sensor_model.{c,h}. Unify on
  (uint8_t*, size_t) in the shared rp2040.h (included by both TUs) and
  drop the divergent declaration + <sys/types.h> from sensor_model.h.
- const-qualifier discard: mqtt.c mqtt_start_interface() dropped the const
  on its parameter (it hands the pointer to a non-const callback that
  mutates the instance).
- unchecked malloc: ha_mqtt.c _mqtt_ha_start() now checks the allocation
  and returns cleanly instead of dereferencing a possible NULL.
- format specifiers: wifi_model.c ping stats use PRIu32 for the uint32_t
  counters instead of %d.
- pointer signedness: wifi_model.c casts uint8_t SSID/password buffers to
  char* at the string-API boundary (wifi_ap_record_t.ssid is uint8_t[]);
  ha_config.c uses size_t for the nvs length out-param.
- dead code / unused symbols: ha_switch.c drops an always-true unsigned
  ">= 0" bound check; wifi_model.c removes the unused _wifi_event_group;
  main.c logs with TAG instead of the literal string "TAG".

Final main/ warning count: 0 (-Wall -Wextra). Remaining build output shows
only ESP-IDF's legacy driver/adc.h deprecation #warning, pulled in via the
vendored bsp_board.h header — not a main/ source warning.

+++ main/ha/ha_config.c
@@ -315,7 +315,7 @@ static void view_event_handler(void *handler_args, esp_event_base_t base, int32_
 
 esp_err_t ha_cfg_get(ha_cfg_interface *ha_cfg)
 {
-    int len = sizeof(ha_cfg_interface);
+    size_t len = sizeof(ha_cfg_interface);
     memset(ha_cfg, 0, sizeof(ha_cfg_interface));
     esp_err_t err = indicator_nvs_read(MQTT_HA_CFG_STORAGE, ha_cfg, &len);
     if (err == ESP_OK && len == sizeof(ha_cfg_interface)) {
